moved the Constants class to the shared module
