+++ shared/src/main/java/com/wizardlybump17/physics/Constants.java
@@ -1,0 +1,11 @@
+package com.wizardlybump17.physics;
+
+public final class Constants {
+
+    public static final int TICKS_PER_SECOND = 128;
+    public static final double MILLIS_PER_TICK = 1000.0 / TICKS_PER_SECOND;
+    public static final double NANOS_PER_TICK = 1_000_000_000.0 / TICKS_PER_SECOND;
+
+    private Constants() {
+    }
+}
