+++ three/src/main/java/com/wizardlybump17/physics/three/object/PhysicsObject.java
@@ -1,6 +1,5 @@
 package com.wizardlybump17.physics.three.object;
 
-import com.wizardlybump17.physics.three.Constants;
 import com.wizardlybump17.physics.three.Vector3D;
 import com.wizardlybump17.physics.three.container.BaseObjectContainer;
 import com.wizardlybump17.physics.three.shape.Shape;
